+++ vector/src/main/java/im/vector/app/features/roomprofile/settings/RoomSettingsViewModel.kt
@@ -60,11 +60,13 @@ class RoomSettingsViewModel @AssistedInject constructor(@Assisted initialState: 
 
     private fun observeState() {
         selectSubscribe(
+                RoomSettingsViewState::avatarAction,
                 RoomSettingsViewState::newName,
                 RoomSettingsViewState::newCanonicalAlias,
                 RoomSettingsViewState::newTopic,
                 RoomSettingsViewState::newHistoryVisibility,
-                RoomSettingsViewState::roomSummary) { newName,
+                RoomSettingsViewState::roomSummary) { avatarAction,
+                                                      newName,
                                                       newCanonicalAlias,
                                                       newTopic,
                                                       newHistoryVisibility,
@@ -72,7 +74,8 @@ class RoomSettingsViewModel @AssistedInject constructor(@Assisted initialState: 
             val summary = asyncSummary()
             setState {
                 copy(
-                        showSaveAction = summary?.name != newName
+                        showSaveAction = avatarAction !is RoomSettingsViewState.AvatarAction.None
+                                || summary?.name != newName
                                 || summary?.topic != newTopic
                                 || summary?.canonicalAlias != newCanonicalAlias?.takeIf { it.isNotEmpty() }
                                 || newHistoryVisibility != null
@@ -101,6 +104,7 @@ class RoomSettingsViewModel @AssistedInject constructor(@Assisted initialState: 
                 .subscribe {
                     val powerLevelsHelper = PowerLevelsHelper(it)
                     val permissions = RoomSettingsViewState.ActionPermissions(
+                            canChangeAvatar = powerLevelsHelper.isUserAllowedToSend(session.myUserId, true, EventType.STATE_ROOM_AVATAR),
                             canChangeName = powerLevelsHelper.isUserAllowedToSend(session.myUserId, true, EventType.STATE_ROOM_NAME),
                             canChangeTopic = powerLevelsHelper.isUserAllowedToSend(session.myUserId, true, EventType.STATE_ROOM_TOPIC),
                             canChangeCanonicalAlias = powerLevelsHelper.isUserAllowedToSend(session.myUserId, true,
@@ -117,6 +121,7 @@ class RoomSettingsViewModel @AssistedInject constructor(@Assisted initialState: 
     override fun handle(action: RoomSettingsAction) {
         when (action) {
             is RoomSettingsAction.EnableEncryption         -> handleEnableEncryption()
+            is RoomSettingsAction.SetAvatarAction          -> setState { copy(avatarAction = action.avatarAction) }
             is RoomSettingsAction.SetRoomName              -> setState { copy(newName = action.newName) }
             is RoomSettingsAction.SetRoomTopic             -> setState { copy(newTopic = action.newTopic) }
             is RoomSettingsAction.SetRoomHistoryVisibility -> setState { copy(newHistoryVisibility = action.visibility) }
@@ -132,6 +137,15 @@ class RoomSettingsViewModel @AssistedInject constructor(@Assisted initialState: 
 
         val summary = state.roomSummary.invoke()
 
+        when (val avatarAction = state.avatarAction) {
+            RoomSettingsViewState.AvatarAction.None            -> Unit
+            RoomSettingsViewState.AvatarAction.DeleteAvatar -> {
+                operationList.add(room.rx().deleteAvatar())
+            }
+            is RoomSettingsViewState.AvatarAction.UpdateAvatar -> {
+                operationList.add(room.rx().updateAvatar(avatarAction.newAvatarUri, avatarAction.newAvatarFileName))
+            }
+        }
         if (summary?.name != state.newName) {
             operationList.add(room.rx().updateName(state.newName ?: ""))
         }